requirements.txt
google-genai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
feedparser>=6.0.10